refactor: split hive flight sql server classes
