+++ examples/tutorial/hive/java-flight-sql-server/src/main/java/org/secretflow/scql/hive/QueryBackend.java
@@ -1,0 +1,14 @@
+package org.secretflow.scql.hive;
+
+import java.sql.SQLException;
+
+interface QueryBackend extends AutoCloseable {
+  String preprocess(String query);
+
+  QueryResources execute(String query) throws SQLException;
+
+  String describe();
+
+  @Override
+  default void close() throws Exception {}
+}
